putting last things with size_t and not int

+++ String.cpp
@@ -98,7 +98,7 @@ bool String::empty(){
 	}
 }
 
-void String::reserve(int n){
+void String::reserve(size_t n){
 	if(n>capacity_){
 		
 	}
